Split into new file
